Izmjenjene postojeće funkcije, optimiziran kod, dodana funkcija za brisanje datoteka, izmjenjen vizualni identitet

+++ Obscure/Form1.cs
@@ -19,7 +19,7 @@ namespace Obscure
 
         private void pictureBox1_Click(object sender, EventArgs e)
         {
-            MessageBox.Show("Odaberite lokaciju na kojoj želite kreirati vaš skriveni folder.", "Obscure", MessageBoxButtons.OK , MessageBoxIcon.Information);
+            MessageBox.Show("Odaberite lokaciju na kojoj želite kreirati vaš skriveni folder ili odaberite postojeći folder.", "Obscure", MessageBoxButtons.OK , MessageBoxIcon.Information);
         }
 
         private void btKreirajFolder_Click(object sender, EventArgs e)
@@ -30,7 +30,6 @@ namespace Obscure
                 StringBuilder ST = new StringBuilder();
 
                 ST.Append(lblLokacija.Text);
-                ST.Append(tbNazivFoldera.Text);
                 ST.Append("\\");
 
                 if (PT.checkThePath(ST.ToString()) == true)
@@ -39,13 +38,12 @@ namespace Obscure
                 }
                 else
                 {
-                    MessageBox.Show("Neispravno kreirana putanja. Koristite ispravno ime i ispravnu putanju.", "Obscure", MessageBoxButtons.OK,MessageBoxIcon.Error);
+                    MessageBox.Show("Neispravna putanja. Odaberite postojeći ili kreirajte novi folder.", "Obscure", MessageBoxButtons.OK,MessageBoxIcon.Error);
                 }
         }
 
         private void btPrikaziSkriveneFoldere_Click(object sender, EventArgs e)
         {
-
             gbKreirajSkriveniFolder.Visible = false;
             gbPopisDatoteka.Visible = false;
             gbPopisSkrivenihFoldera.Visible = true;
@@ -60,11 +58,19 @@ namespace Obscure
 
         private void btPrikaziSkriveniFolder_Click(object sender, EventArgs e)
         {
-            foreach (DataGridViewCell cell in dgView1.SelectedCells)
+            if (gbPopisSkrivenihFoldera.Visible == true)
             {
-                ShowFolder SH = new ShowFolder();
-                SH.showFolder(cell.Value.ToString());   
-                MessageBox.Show(String.Format("Folder: {0} više nije sakriven.", cell.Value.ToString()), "Obscure", MessageBoxButtons.OK, MessageBoxIcon.Information);             
+                foreach (DataGridViewCell cell in dgView1.SelectedCells)
+                {
+                    ShowFolder SH = new ShowFolder();
+                    SH.showFolder(cell.Value.ToString());
+                    MessageBox.Show(String.Format("Folder: {0} više nije sakriven.", cell.Value.ToString()), "Obscure", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    btnPretraži_Click(sender, e);
+                }
+            }
+            else
+            {
+                btPrikaziSkriveneFoldere_Click(sender, e);
             }
         }
 
@@ -74,7 +80,7 @@ namespace Obscure
             lblLokacija.Text = FB.folderBrowser();
         }
 
-        public void btnLokacijaPretrage_Click(object sender, EventArgs e)
+        private void btnLokacijaPretrage_Click(object sender, EventArgs e)
         {
             FolderBrowser FB = new FolderBrowser();
             lblLokacijaSkrivenihFoldera.Text = FB.folderBrowser(); ;
@@ -82,17 +88,24 @@ namespace Obscure
 
         private void btnPretraži_Click(object sender, EventArgs e)
         {
-            SearchHiddenFolders S = new SearchHiddenFolders();
-
-            if (S.searchForHiddenFolders(lblLokacijaSkrivenihFoldera.Text).Count() == 0)
-            {
-                MessageBox.Show("Na navedenoj lokaciji ne postoji niti jedan skriveni folder", "Obscure", MessageBoxButtons.OK);
-            }
+            if (lblLokacijaSkrivenihFoldera.Text == "")
+                MessageBox.Show("Niste odabrali lokaciju pretrage.", "Obscure", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             else
             {
-                FillDataGrid FD = new FillDataGrid();
-                FD.fillDataGridWithData(S.searchForHiddenFolders(lblLokacijaSkrivenihFoldera.Text), dgView1);
+                SearchHiddenFolders S = new SearchHiddenFolders();
 
+                if (S.searchForHiddenFolders(lblLokacijaSkrivenihFoldera.Text).Count() == 0)
+                {
+                    MessageBox.Show("Na navedenoj lokaciji ne postoji niti jedan skriveni folder", "Obscure", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    dgView1.DataSource = null;
+                    lblNum1.Text = "0";
+                }
+                else
+                {
+                    FillDataGrid FD = new FillDataGrid();
+                    FD.fillDataGridWithData(S.searchForHiddenFolders(lblLokacijaSkrivenihFoldera.Text), dgView1);
+
+                }
             }
         }
 
@@ -100,10 +113,17 @@ namespace Obscure
         {
             DeleteFolder D = new DeleteFolder();
 
-            foreach (DataGridViewCell cell in dgView1.SelectedCells)
+            if (gbPopisSkrivenihFoldera.Visible == true)
             {
-                D.deleteFolder(cell.Value.ToString());
-                btnPretraži_Click(sender, e);
+                foreach (DataGridViewCell cell in dgView1.SelectedCells)
+                {
+                    D.deleteFolder(cell.Value.ToString());
+                    btnPretraži_Click(sender, e);
+                }
+            }
+            else
+            {
+                btPrikaziSkriveneFoldere_Click(sender, e);
             }
         }
 
@@ -120,9 +140,26 @@ namespace Obscure
 
             foreach (DataGridViewCell cell in dgView1.SelectedCells)
             {
-                FD.fillDataGridWithData(SF.showFiles(cell.Value.ToString(), files), dgView2);
+                if (SF.showFiles(cell.Value.ToString(), files).Count() == 0)
+                {
+                    MessageBox.Show("Odabrani folder ne sadrži niti jednu datoteku.", "Obscure", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+                else
+                {
+                    FD.fillDataGridWithData(SF.showFiles(cell.Value.ToString(), files), dgView2);
+                }
             }
 
+        }
+
+        private void dgView1_DataBindingComplete(object sender, DataGridViewBindingCompleteEventArgs e)
+        {
+                lblNum1.Text = dgView1.RowCount.ToString();
+        }
+
+        private void dgView2_DataBindingComplete(object sender, DataGridViewBindingCompleteEventArgs e)
+        {
+                lblNum2.Text = dgView1.RowCount.ToString();
         }
     }
 }
